correction de la methode update en modifiant le type de user id(long en ----->int)

+++ ProjetGestiBanck/GestiBankBack/src/main/java/com/wha/springmvc/model/User.java
@@ -14,7 +14,7 @@ public class User {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private long id;
+	private int id;
 	
 	private String username;
 	
@@ -26,18 +26,18 @@ public class User {
 		id=0;
 	}
 	
-	public User(long id, String username, String address, String email){
+	public User(int id, String username, String address, String email){
 		this.id = id;
 		this.username = username;
 		this.address = address;
 		this.email = email;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
